requirements.txt
crewai==0.134.0
crewai-tools==0.48.0
langgraph==0.4.8
python-dotenv==1.1.0
google_search_results==2.4.2
google-api-python-client==2.173.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
langchain-community~=0.3.26
langchain~=0.3.26
protobuf~=5.29.5